+++ build.sh
@@ -4,10 +4,12 @@ echo "Building Tumblr Savior $VERSION"
 
 cd src
 zip ../Tumblr-Savior-$VERSION.zip ./* -r -9
-cfx xpi --output-file Tumblr-Savior-$VERISON.xpi
+cfx xpi --output-file ../Tumblr-Savior-$VERSION.xpi
+
+cd ..
 
 mv src src.safariextension
 ./build-safari-extension.sh src.safariextension
 mv src.safariextension src
 
-mv src.safariextz Tumblr-Savior-$VERSION.safarietz
+mv src.safariextz Tumblr-Savior-$VERSION.safariextz